Finished reading
Fragments, climate change. From New Scientists review "Offill’s Weather makes one consider the inevitable and accelerating encroachment of climate change on the foundations of what we consider normal life. It is a reminder that in a time of crisis, when climate emergency threatens to worsen into catastrophe, inaction is a moral failure."